Completed Custom ObsRaw
